Add Lookup data and class
Lookup data is domain-specific semi-static cached data the application
uses while executing, without accessing the database. It's mainly used
to map IDs coming from the database to names and labels, without the
need of a JOIN query, since there are many IDs but a very limited
amount of possible values.