Refactor helper functions
Helpers functions are grouped and indexed at the beginning of the file.
Functions with a *_old() equivalent currently return the *_old()
definition but this is needed to make room for improved version of said
functions. Next step is to change all these function calls in the code
to the *_old() definition. Ex.: redirect() => redirect_old()