Add base classes CrudService and InputProcessor
CrudService is an abstract class which provides a CrudServiceInterface
for abstracting business logic and thus managing the persistence layer.

Simply put, every CRUD (or, better, CUD aka Create, Update and Delete)
action maps to a single CrudService child, e.g. RulingCreateService,
and exposes some simple API like processInput() to process POST input,
syncDatabase() and syncFilesystem() to perform operations on those
persistence layers (e.g. storing new data into the database or an image
file) and getFeedback(), which returns an array like [message, uri] to
redirect to that URI and then show that alert message to the user.

A CrudService, if given input and an input processor class name,
instantiates an InputProcessor child class, like RulingInputProcessor

An InputProcessor is solely responsibile to handle already-validated
input and return an associative array of processed data that *SHOULD*
match database columns or at least part of them.